# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
        "/categories": {
            "get": {
                "description": "Get all categories",
                "produces": [
                    "application/json"
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/entity.CategoriesResponse"
                        }
                    }
                }
            }
        },
        "/categories/:id": {
            "delete": {
                "description": "Delete a category. Must be contained in no dreams.",
                "produces": [
                    "application/json"
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
        "/categories/:id/name": {
            "patch": {
                "description": "Change the name of a category.",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "type": "string",
                        "description": "New name for this category",
                        "name": "name",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
        "/categories/:id/type": {
            "patch": {
                "description": "Change the type of a category.",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "type": "string",
                        "description": "New type for this category",
                        "name": "type",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
        "/categories/merge": {
            "post": {
                "description": "Merge two categories.",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "description": "Which categories should be merged",
                        "name": "mergeCategoriesParams",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/controller.MergeCategoriesParams"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/entity.CategoriesCountResponse"
                        }
                    }
                }
            }
        },
        "/count-categories": {
            "get": {
                "description": "Get count per category and person",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "type": "number",
                        "description": "Limit of returned results",
                        "name": "limit",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Counts by category and persons",
                        "schema": {
                            "$ref": "#/definitions/entity.CategoriesCountResponse"
                        }
                    }
                }
            }
        },
        "/count-categories/monthly": {
            "get": {
                "description": "Get count per month",
                "produces": [
                    "application/json"
                ],
                "responses": {
                    "200": {
                        "description": "Monthly statistics",
                        "schema": {
                            "$ref": "#/definitions/entity.Statistics"
                        }
                    }
                }
            }
        },
        "/dreams": {
            "get": {
                "description": "Get all dreams",
                "produces": [
                    "application/json"
                ],
                "responses": {
                    "200": {
                        "description": "List of all dreams",
                        "schema": {
                            "$ref": "#/definitions/entity.DreamsResponse"
                        }
                    }
                }
            },
            "post": {
                "description": "Create a new dream",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "description": "The dream which will be created",
                        "name": "postDreamRequest",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/controller.PostDreamRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "type": "number"
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            }
        },
        "/dreams/private": {
            "get": {
                "security": [
                    {
                        "BasicAuth": []
                    }
                ],
                "description": "Get all dreams - including private",
                "produces": [
                    "application/json"
                ],
                "responses": {
                    "200": {
                        "description": "List of all dreams",
                        "schema": {
                            "$ref": "#/definitions/entity.DreamsResponse"
                        }
                    }
                }
            }
        },
        "/dreams/private/{dreamId}": {
            "get": {
                "security": [
                    {
                        "BasicAuth": []
                    }
                ],
                "description": "Get one private dream",
                "produces": [
                    "application/json"
                ],
                "responses": {
                    "200": {
                        "description": "One dream",
                        "schema": {
                            "$ref": "#/definitions/entity.DreamResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "404": {
                        "description": "Not Found"
                    }
                }
            },
            "patch": {
                "security": [
                    {
                        "BasicAuth": []
                    }
                ],
                "description": "Toggle visibility of a dream",
                "produces": [
                    "application/json"
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
        "/dreams/{dreamId}": {
            "get": {
                "description": "Get one dream",
                "produces": [
                    "application/json"
                ],
                "responses": {
                    "200": {
                        "description": "One dream",
                        "schema": {
                            "$ref": "#/definitions/entity.DreamResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "404": {
                        "description": "Not Found"
                    }
                }
            },
            "delete": {
                "description": "Delete one dreams",
                "produces": [
                    "application/json"
                ],
                "responses": {
                    "400": {
                        "description": "Bad Request"
                    },
                    "404": {
                        "description": "Not Found"
                    }
                }
            },
            "patch": {
                "description": "Update an existing dream",
                "consumes": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "description": "All parameters of the dream that should be updated",
                        "name": "updateDreamRequest",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/controller.UpdateDreamRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "404": {
                        "description": "Not Found"
                    }
                }
            }
        },
        "/dreams/{dreamId}/categories": {
            "post": {
                "description": "Add a new person or category to a dream",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "description": "Name and category. Allowed values for 'catgory' are: 'person', 'category'",
                        "name": "params",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/controller.PostCategoryRequestBody"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "type": "number"
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "404": {
                        "description": "Not Found"
                    }
                }
            }
        },
        "/dreams/{dreamId}/categories/{categoryId}": {
            "put": {
                "description": "Add an exiting person or category to a dream",
                "produces": [
                    "application/json"
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "404": {
                        "description": "Not Found"
                    }
                }
            },
            "delete": {
                "description": "Delete a category or person from a dream",
                "produces": [
                    "application/json"
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "404": {
                        "description": "Not Found"
                    }
                }
            }
        },
        "/dreams/{dreamId}/finalize": {
            "patch": {
                "description": "Finalize a dream",
                "produces": [
                    "application/json"
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "404": {
                        "description": "Not Found"
                    }
                }
            }
        },
        "/login": {
            "post": {
                "description": "Login",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "description": "Password and user name",
                        "name": "loginRequest",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/controller.LoginRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/entity.LoginResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    }
                }
            }
        },
        "/logout": {
            "post": {
                "description": "Logout",
                "produces": [
                    "application/json"
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            }
        }
    },
    "definitions": {
        "controller.LoginRequest": {
            "type": "object",
            "required": [
                "name",
                "password"
            ],
            "properties": {
                "name": {
                    "type": "string"
                },
                "password": {
                    "type": "string"
                }
            }
        },
        "controller.MergeCategoriesParams": {
            "type": "object",
            "required": [
                "newName",
                "sourceCategoryId",
                "targetCategoryId"
            ],
            "properties": {
                "newName": {
                    "type": "string"
                },
                "sourceCategoryId": {
                    "type": "integer"
                },
                "targetCategoryId": {
                    "type": "integer"
                }
            }
        },
        "controller.PostCategoryRequestBody": {
            "type": "object",
            "required": [
                "categoryType",
                "name"
            ],
            "properties": {
                "categoryType": {
                    "$ref": "#/definitions/entity.CategoryType"
                },
                "name": {
                    "type": "string"
                }
            }
        },
        "controller.PostDreamRequest": {
            "type": "object",
            "required": [
                "date"
            ],
            "properties": {
                "date": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                }
            }
        },
        "controller.UpdateDreamRequest": {
            "type": "object",
            "properties": {
                "date": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "rating": {
                    "type": "integer"
                }
            }
        },
        "entity.CategoriesCountResponse": {
            "type": "object",
            "required": [
                "categories"
            ],
            "properties": {
                "categories": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/entity.CountByCat"
                    }
                }
            }
        },
        "entity.CategoriesResponse": {
            "type": "object",
            "required": [
                "categories"
            ],
            "properties": {
                "categories": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/entity.CategoryResponse"
                    }
                }
            }
        },
        "entity.CategoryResponse": {
            "type": "object",
            "required": [
                "id",
                "name",
                "type"
            ],
            "properties": {
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "type": {
                    "$ref": "#/definitions/entity.CategoryType"
                }
            }
        },
        "entity.CategoryType": {
            "type": "string",
            "enum": [
                "person",
                "category"
            ],
            "x-enum-varnames": [
                "TypePerson",
                "TypeCategory"
            ]
        },
        "entity.CountByCat": {
            "type": "object",
            "required": [
                "count",
                "id"
            ],
            "properties": {
                "count": {
                    "type": "integer"
                },
                "id": {
                    "type": "integer"
                }
            }
        },
        "entity.DreamMetaResponse": {
            "type": "object",
            "required": [
                "date",
                "finalized",
                "id",
                "visible"
            ],
            "properties": {
                "date": {
                    "type": "string"
                },
                "finalized": {
                    "type": "boolean"
                },
                "id": {
                    "type": "integer"
                },
                "rating": {
                    "type": "integer"
                },
                "visible": {
                    "type": "boolean"
                }
            }
        },
        "entity.DreamResponse": {
            "type": "object",
            "required": [
                "categories",
                "date",
                "description",
                "finalized",
                "id",
                "visible"
            ],
            "properties": {
                "categories": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/entity.CategoryResponse"
                    }
                },
                "date": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "finalized": {
                    "type": "boolean"
                },
                "id": {
                    "type": "integer"
                },
                "rating": {
                    "type": "integer"
                },
                "visible": {
                    "type": "boolean"
                }
            }
        },
        "entity.DreamsResponse": {
            "type": "object",
            "required": [
                "dreams"
            ],
            "properties": {
                "dreams": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/entity.DreamMetaResponse"
                    }
                }
            }
        },
        "entity.LoginResponse": {
            "type": "object",
            "required": [
                "token"
            ],
            "properties": {
                "token": {
                    "type": "string"
                }
            }
        },
        "entity.Statistic": {
            "type": "object",
            "required": [
                "categories",
                "dreamCount",
                "month"
            ],
            "properties": {
                "categories": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/entity.CountByCat"
                    }
                },
                "dreamCount": {
                    "type": "integer"
                },
                "month": {
                    "type": "string"
                }
            }
        },
        "entity.Statistics": {
            "type": "object",
            "required": [
                "statistics"
            ],
            "properties": {
                "statistics": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/entity.Statistic"
                    }
                }
            }
        }
    }
}